refactor the constants
